refactor: simplify cpu stepping and memory map constants
