+++ src/bus.rs
@@ -1,15 +1,15 @@
 use crate::cartridge::Cartridge;
-
-const WRAM_SIZE: usize = 0x2000;
-const OAM_SIZE: usize = 0xA0;
-const IO_SIZE: usize = 0x80;
-const HRAM_SIZE: usize = 0x7F;
+use crate::memory_map::{
+    CARTRIDGE_ROM_END, CARTRIDGE_ROM_START, ECHO_RAM_END, ECHO_RAM_START, HRAM_END, HRAM_SIZE,
+    HRAM_START, INTERRUPT_ENABLE, IO_REGISTERS_END, IO_REGISTERS_SIZE, IO_REGISTERS_START, OAM_END,
+    OAM_SIZE, OAM_START, UNUSABLE_END, UNUSABLE_START, WRAM_END, WRAM_SIZE, WRAM_START,
+};
 
 pub struct Bus {
     cartridge: Cartridge,
     wram: [u8; WRAM_SIZE],
     oam: [u8; OAM_SIZE],
-    io_registers: [u8; IO_SIZE],
+    io_registers: [u8; IO_REGISTERS_SIZE],
     hram: [u8; HRAM_SIZE],
     interrupt_enable: u8,
 }
@@ -20,7 +20,7 @@ impl Bus {
             cartridge,
             wram: [0; WRAM_SIZE],
             oam: [0; OAM_SIZE],
-            io_registers: [0; IO_SIZE],
+            io_registers: [0; IO_REGISTERS_SIZE],
             hram: [0; HRAM_SIZE],
             interrupt_enable: 0,
         }
@@ -28,51 +28,51 @@ impl Bus {
 
     pub fn read8(&self, addr: u16) -> u8 {
         match addr {
-            0x0000..=0x7FFF => self.cartridge.read_rom(addr),
-            0xC000..=0xDFFF => self.wram[wram_index(addr)],
-            0xE000..=0xFDFF => self.wram[echo_wram_index(addr)],
-            0xFE00..=0xFE9F => self.oam[oam_index(addr)],
-            0xFEA0..=0xFEFF => 0xFF,
-            0xFF00..=0xFF7F => self.io_registers[io_index(addr)],
-            0xFF80..=0xFFFE => self.hram[hram_index(addr)],
-            0xFFFF => self.interrupt_enable,
+            CARTRIDGE_ROM_START..=CARTRIDGE_ROM_END => self.cartridge.read_rom(addr),
+            WRAM_START..=WRAM_END => self.wram[wram_index(addr)],
+            ECHO_RAM_START..=ECHO_RAM_END => self.wram[echo_wram_index(addr)],
+            OAM_START..=OAM_END => self.oam[oam_index(addr)],
+            UNUSABLE_START..=UNUSABLE_END => 0xFF,
+            IO_REGISTERS_START..=IO_REGISTERS_END => self.io_registers[io_index(addr)],
+            HRAM_START..=HRAM_END => self.hram[hram_index(addr)],
+            INTERRUPT_ENABLE => self.interrupt_enable,
             _ => 0xFF,
         }
     }
 
     pub fn write8(&mut self, addr: u16, value: u8) {
         match addr {
-            0x0000..=0x7FFF => self.cartridge.write_rom(addr, value),
-            0xC000..=0xDFFF => self.wram[wram_index(addr)] = value,
-            0xE000..=0xFDFF => self.wram[echo_wram_index(addr)] = value,
-            0xFE00..=0xFE9F => self.oam[oam_index(addr)] = value,
-            0xFEA0..=0xFEFF => {}
-            0xFF00..=0xFF7F => self.io_registers[io_index(addr)] = value,
-            0xFF80..=0xFFFE => self.hram[hram_index(addr)] = value,
-            0xFFFF => self.interrupt_enable = value,
+            CARTRIDGE_ROM_START..=CARTRIDGE_ROM_END => self.cartridge.write_rom(addr, value),
+            WRAM_START..=WRAM_END => self.wram[wram_index(addr)] = value,
+            ECHO_RAM_START..=ECHO_RAM_END => self.wram[echo_wram_index(addr)] = value,
+            OAM_START..=OAM_END => self.oam[oam_index(addr)] = value,
+            UNUSABLE_START..=UNUSABLE_END => {}
+            IO_REGISTERS_START..=IO_REGISTERS_END => self.io_registers[io_index(addr)] = value,
+            HRAM_START..=HRAM_END => self.hram[hram_index(addr)] = value,
+            INTERRUPT_ENABLE => self.interrupt_enable = value,
             _ => {}
         }
     }
 }
 
 fn wram_index(addr: u16) -> usize {
-    usize::from(addr - 0xC000)
+    usize::from(addr - WRAM_START)
 }
 
 fn echo_wram_index(addr: u16) -> usize {
-    usize::from(addr - 0xE000) % WRAM_SIZE
+    usize::from(addr - ECHO_RAM_START) % WRAM_SIZE
 }
 
 fn oam_index(addr: u16) -> usize {
-    usize::from(addr - 0xFE00)
+    usize::from(addr - OAM_START)
 }
 
 fn io_index(addr: u16) -> usize {
-    usize::from(addr - 0xFF00)
+    usize::from(addr - IO_REGISTERS_START)
 }
 
 fn hram_index(addr: u16) -> usize {
-    usize::from(addr - 0xFF80)
+    usize::from(addr - HRAM_START)
 }
 
 #[cfg(test)]
@@ -80,9 +80,9 @@ mod tests {
     use super::*;
 
     fn bus() -> Bus {
-        let mut rom = vec![0; 0x8000];
-        rom[0x0000] = 0xAA;
-        rom[0x7FFF] = 0xBB;
+        let mut rom = vec![0; crate::memory_map::CARTRIDGE_ROM_SIZE];
+        rom[usize::from(CARTRIDGE_ROM_START)] = 0xAA;
+        rom[usize::from(CARTRIDGE_ROM_END)] = 0xBB;
         Bus::new(Cartridge::from_bytes(rom).expect("valid ROM-only cartridge"))
     }
 
@@ -90,21 +90,21 @@ mod tests {
     fn delegates_rom_reads_and_writes_to_cartridge() {
         let mut bus = bus();
 
-        assert_eq!(bus.read8(0x0000), 0xAA);
-        assert_eq!(bus.read8(0x7FFF), 0xBB);
-        bus.write8(0x0000, 0x99);
-        assert_eq!(bus.read8(0x0000), 0xAA);
+        assert_eq!(bus.read8(CARTRIDGE_ROM_START), 0xAA);
+        assert_eq!(bus.read8(CARTRIDGE_ROM_END), 0xBB);
+        bus.write8(CARTRIDGE_ROM_START, 0x99);
+        assert_eq!(bus.read8(CARTRIDGE_ROM_START), 0xAA);
     }
 
     #[test]
     fn mirrors_wram_and_echo_ram() {
         let mut bus = bus();
 
-        bus.write8(0xC000, 0x12);
+        bus.write8(WRAM_START, 0x12);
         assert_eq!(bus.read8(0xE000), 0x12);
 
         bus.write8(0xE000, 0x34);
-        assert_eq!(bus.read8(0xC000), 0x34);
+        assert_eq!(bus.read8(WRAM_START), 0x34);
     }
 
     #[test]
@@ -133,23 +133,23 @@ mod tests {
     fn io_hram_and_ie_are_writable() {
         let mut bus = bus();
 
-        bus.write8(0xFF00, 0x12);
+        bus.write8(IO_REGISTERS_START, 0x12);
         bus.write8(0xFF80, 0x34);
-        bus.write8(0xFFFF, 0x56);
+        bus.write8(INTERRUPT_ENABLE, 0x56);
 
-        assert_eq!(bus.read8(0xFF00), 0x12);
+        assert_eq!(bus.read8(IO_REGISTERS_START), 0x12);
         assert_eq!(bus.read8(0xFF80), 0x34);
-        assert_eq!(bus.read8(0xFFFF), 0x56);
+        assert_eq!(bus.read8(INTERRUPT_ENABLE), 0x56);
     }
 
     #[test]
     fn unmapped_ranges_read_ff_and_ignore_writes() {
         let mut bus = bus();
 
-        bus.write8(0x8000, 0x12);
+        bus.write8(crate::memory_map::VRAM_START, 0x12);
         bus.write8(0xA000, 0x34);
 
-        assert_eq!(bus.read8(0x8000), 0xFF);
+        assert_eq!(bus.read8(crate::memory_map::VRAM_START), 0xFF);
         assert_eq!(bus.read8(0xA000), 0xFF);
     }
 }
