+++ src/lib.rs
@@ -1,3 +1,4 @@
 pub mod bus;
 pub mod cartridge;
 pub mod cpu;
+pub mod memory_map;
